Task tweak to swallow errors

+++ tasks/pack.py
@@ -10,6 +10,7 @@ ROOT = pathlib.Path(__file__).resolve().parent.parent
 @invoke.task()
 def build(ctx):
     """Build the package into distributables.
+
     This will create two distributables: source and wheel.
     """
     ctx.run(f'python setup.py sdist bdist_wheel')
@@ -22,13 +23,16 @@ def clean(ctx):
     ctx.run(f'python setup.py clean')
     dist = ROOT.joinpath('dist')
     print(f'[pack:clean] Removing {dist}')
-    shutil.rmtree(str(dist))
+    if dist.exists():
+        shutil.rmtree(str(dist))
 
 
 @invoke.task(pre=[clean, build])
 def upload(ctx, repo):
     """Upload the package to an index server.
+
     This implies cleaning and re-building the package.
+
     :param repo: Required. Name of the index server to upload to, as specifies
         in your .pypirc configuration file.
     """
